Do not materialiaze soft deleted assets in server responses
