+++ src/database/model/ai_resource/resource_table.py
@@ -13,13 +13,17 @@ class AIResourcePartLink(SQLModel, table=True):  # type: ignore [call-arg]
     __tablename__ = "ai_resource_part_link"
     parent_identifier: str = Field(
         max_length=IDENTIFIER_LENGTH,
-        sa_column_args=[ForeignKey("ai_resource.identifier", onupdate="CASCADE")],
+        sa_column_args=[
+            ForeignKey("ai_resource.identifier", onupdate="CASCADE", ondelete="CASCADE")
+        ],
         sa_column_kwargs=dict(nullable=True, index=True),
         primary_key=True,
     )
     child_identifier: str = Field(
         max_length=IDENTIFIER_LENGTH,
-        sa_column_args=[ForeignKey("ai_resource.identifier", onupdate="CASCADE")],
+        sa_column_args=[
+            ForeignKey("ai_resource.identifier", onupdate="CASCADE", ondelete="CASCADE")
+        ],
         sa_column_kwargs=dict(nullable=True, index=True),
         primary_key=True,
     )
@@ -29,13 +33,17 @@ class AIResourceRelevantLink(SQLModel, table=True):  # type: ignore [call-arg]
     __tablename__ = "ai_resource_relevant_link"
     parent_identifier: str = Field(
         max_length=IDENTIFIER_LENGTH,
-        sa_column_args=[ForeignKey("ai_resource.identifier", onupdate="CASCADE")],
+        sa_column_args=[
+            ForeignKey("ai_resource.identifier", onupdate="CASCADE", ondelete="CASCADE")
+        ],
         sa_column_kwargs=dict(nullable=True, index=True),
         primary_key=True,
     )
     relevant_identifier: str = Field(
         max_length=IDENTIFIER_LENGTH,
-        sa_column_args=[ForeignKey("ai_resource.identifier", onupdate="CASCADE")],
+        sa_column_args=[
+            ForeignKey("ai_resource.identifier", onupdate="CASCADE", ondelete="CASCADE")
+        ],
         sa_column_kwargs=dict(nullable=True, index=True),
         primary_key=True,
     )
