+++ src/tests/routers/resource_routers/test_router_resource_bundle.py
@@ -4,32 +4,26 @@ from unittest.mock import Mock
 from starlette.testclient import TestClient
 from database.model.ai_resource.resource_table import AIResourceORM
 from database.session import DbSession
+from tests.testutils.users import register_asset, ALICE
 
 
 def test_resource_bundle_api(
     client: TestClient,
     mocked_privileged_token: Mock,
+    publication_factory,
     body_asset: dict,
     auto_publish: None,
 ):
     """
     Test creating and retrieving a ResourceBundle through the API.
     """
-    resource_1 = AIResourceORM(name="AI Resource 1")
-    resource_2 = AIResourceORM(name="AI Resource 2")
-
-
-    with DbSession() as session:
-        session.add(resource_1)
-        session.add(resource_2)
-        session.commit()
-        session.refresh(resource_1)
-        session.refresh(resource_2)
+    resource_1 = register_asset(publication_factory(), owner=ALICE)
+    resource_2 = register_asset(publication_factory(), owner=ALICE)
 
     # Prepare request body
     body = copy.copy(body_asset)
     body["name"] = "My AI Bundle"
-    body["includes_resource"] = [resource_1.identifier, resource_2.identifier]
+    body["includes_resource"] = [resource_1, resource_2]
     body["includes_external_reference"] = ["https://example.com/resource"]
 
 
@@ -42,5 +36,5 @@ def test_resource_bundle_api(
 
     response_json = response.json()
     assert response_json["name"] == "My AI Bundle"
-    assert set(response_json["includes_resource"]) == {resource_1.identifier, resource_2.identifier}
+    assert set(response_json["includes_resource"]) == {resource_1, resource_2}
     assert response_json["includes_external_reference"] == ["https://example.com/resource"]
